Replace minimal_transformer_training by a truly minimal example
The existing example is moved to custom_transformer_training.py.
