+++ examples/library/README.md
@@ -2,4 +2,5 @@
 
 This directory contains examples using OpenNMT-tf as a library. This can be useful to run custom experiments without the need to dive into the toolkit.
 
-* `minimal_transformer_training.py` minimal Transformer training and translation script.
+* `minimal_transformer_training.py`: minimal Transformer training and translation script using high-level APIs.
+* `custom_transformer_training.py`: custom training and translation loop demonstrating dataset, encoder, and decoder APIs.
